Add RSA signature protection for DHT records (#187)
This PR introduces a notion of protected DHT records whose key/subkey contains substring [owner:ssh-rsa ...] (the format can be changed) with an RSA public key of the owner. Changes to such records always must be signed with the corresponding private key (so only the owner can change them). This protects from malicious nodes trying to spoil the DHT contents.