Remove use of packaging module (#284)
Currently, setup.py uses packaging module to parse Golang versions. This leads to errors in case of installing hivemind to a clean Python installation, since packaging is not a part of setuptools and is not installed by default.

This PR solves the problem by removing the usage of the packaging module completely.