Improve user-friendliness and fix misc errors (#428)
This PR introduces two quality-of-life changes found with the help of community members

1. Warn if there are too many no-grad params (Found by mr_seeker@ from [KoboldAI](https://github.com/KoboldAI))

The problem manifested itself when he created an optimizer over GPT-J-6B with most parameters set to requires_grad_(False).

Currently, TrainingStateAverager will still average these parameters with peers to properly average batchnorm EMAs and similar statistics. As a result, TrainingStateAverager would OOM after trying to initialize buffers for all gpt-j parameters.

This PR checks if the majority of parameters are non-trainable and prints a warning with instructions.

2. Do not count metric loggers as training peers (found with the help of CALM collaborators)

Currently, non-gpu peers that are responsible for metric logging periodically report having accumulated 0 samples at eps samples/sec. This causes confusion because these peers do not actually compute minibatches and are NOT counted in "averaged gradients with X peers".

3. found in @bawr 's logs
Optimizer with DPU and HivemindGradScaler would trigger an error if gradient overflow was detected during DPU

4. found by @pr-mais : setting ANNOUNCE_MADDRS with port 0 would silently break matchmaking with some probability
The reason behind this is that P2P does not auto-resolve port 0 in announce_maddrs. We will now raise an error if user sets up a P2P with this configuration error

5. found by @borzunov  - Reducer error when handling network error
This is a non-fatal error that happens on reducer when some senders take too long to send their data and are killed -- but their last message does make it.



Co-authored-by: Mais Alheraki <mais@fairybits.com>
Co-authored-by: Alexander Borzunov <hxrussia@gmail.com>
Co-authored-by: Julius ter Pelkwijk <1099127+mrseeker@users.noreply.github.com>
Co-authored-by: Max Ryabinin <mryabinin0@gmail.com>