Update p2pd to v0.3.8 (and libp2p to v0.17.0) (#469)
This PR includes refactoring of `setup.py` making it easier to include more binaries into the distribution. The current version of `p2p-keygen` can be easily reimplemented in Python, so we don't actually include it now (but keep the opportunity to do that in case if this util becomes more sophisticated).