Fix p2pd cross-compilation, build wheels in GHA (#665)
It was discovered that issues such as #658 and #655 were caused by an accidental publish of a wheel that contained a platform-specific p2pd binary. To prevent such issues from happening in the future, this PR enables broader cross-compilation support and adds a CI action to publish built wheels to PyPI once the release is made.

Fixes #658 and #655